Make statemgr independent of test packages.
